fix(ActivityTracking): Adjust time unit from milliseconds to seconds
Updated voice activity duration calculations and related code to use seconds instead of milliseconds for consistency and improved readability.

+++ Asunara-Event-Manager/Data/Repositories/Models/ActivityTopResult.cs
@@ -4,5 +4,5 @@ public class ActivityTopResult
 {
     public ulong DiscordUserId { get; set; }
     
-    public int Count { get; set; }
+    public long Count { get; set; }
 }
